+++ VeniVidiDaVinci/src/GameScreenMainMenu.h
@@ -32,6 +32,6 @@ public:
 	~GameScreenMainMenu();
 
 	void Render();
-	void Update(float deltaTime, SDL_Event event);
+	void Update(double deltaTime, SDL_Event event);
 };
